# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Shop System API",
        "description": "API documentation for the Online Shop System. Supports public, customer, and admin endpoints.",
        "contact": {
            "email": "support@onlineshop.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/admin/orders/stats": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get order statistics",
                "operationId": "fb49dc1a775d8f5ae2177a6a5966ffc6",
                "responses": {
                    "200": {
                        "description": "Order stats"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/orders/earnings": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get earnings data for charts",
                "operationId": "7f65116c205e6e5c6bc80b092101fa4f",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Period (daily, weekly, monthly)",
                        "schema": {
                            "type": "string",
                            "default": "monthly"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Earnings data with labels and values"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List orders for authenticated user",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Paginated list of orders"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter your Bearer token obtained from login/register. Format: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}